OLH-1663 - Update the frontend state machine for adding an authenticator app
