+++ src/components/add-mfa-method-app/index.njk
@@ -4,6 +4,8 @@
 {% from "govuk/components/details/macro.njk" import govukDetails %}
 {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
 
+{% set pageTitleName = 'pages.addMfaMethod.app.title' | translate %}
+
 {% block pageContent %}
     {{ govukErrorSummary({
         errorList: errorList,
