+++ src/components/add-mfa-method/index.njk
@@ -3,7 +3,7 @@
 {% from "govuk/components/details/macro.njk" import govukDetails %}
 {% from "govuk/components/button/macro.njk" import govukButton %}
 
-
+{% set pageTitleName = 'pages.addMfaMethod.title' | translate %}
 {% block pageContent %}
     <form action="{{ "ADD_MFA_METHOD" | getPath }}" method="post">
     <input type="hidden" name="_csrf" value="{{ csrfToken }}">
