+++ src/components/add-mfa-method/tests/add-mfa-methods-controller.test.ts
@@ -1,19 +1,52 @@
 import { expect } from "chai";
 import { describe, it } from "mocha";
 import { Request, Response } from "express";
-import sinon, { SinonSandbox } from "sinon";
+import { SinonSandbox } from "sinon";
+import { sinon } from "../../../../test/utils/test-utils";
+
 import {
   addMfaMethodPost,
-  addMfaAppMethodGet,
-  addMfaAppMethodPost,
+  addMfaMethodGet,
 } from "../add-mfa-methods-controller";
 import { PATH_DATA } from "../../../app.constants";
-import * as mfaModule from "../../../utils/mfa";
-import QRCode from "qrcode";
+
+describe("addMfaMethodGet", () => {
+  let sandbox: sinon.SinonSandbox;
+  let req: Partial<Request>;
+  let res: Partial<Response>;
+
+  beforeEach(() => {
+    sandbox = sinon.createSandbox();
+
+    req = {
+      body: {},
+      session: { user: { state: { addMfaMethod: ["AUTHENTICATED"] } } } as any,
+      cookies: { lng: "en" },
+      i18n: { language: "en" },
+      t: sandbox.fake(),
+    };
+    res = {
+      render: sandbox.fake(),
+      redirect: sandbox.fake(() => {}),
+      locals: {},
+      status: sandbox.fake(),
+    };
+  });
+
+  afterEach(() => {
+    sandbox.restore();
+  });
+
+  it("should add mfa method page", () => {
+    addMfaMethodGet(req as Request, res as Response);
+
+    expect(res.render).to.have.calledWith("add-mfa-method/index.njk");
+  });
+});
 
 describe("addMfaMethodPost", () => {
   let sandbox: SinonSandbox;
-  let req: Partial<Request>;
+  let req: any;
   let res: Partial<Response>;
   let next: sinon.SinonSpy;
 
@@ -34,7 +67,11 @@ describe("addMfaMethodPost", () => {
         return {} as Response; // You should return a proper type here, matching your test environment needs
       }
     );
-    req = { body: {}, log: { error: sandbox.fake() } as any };
+    req = {
+      body: {},
+      log: { error: sandbox.fake() } as any,
+      session: { user: { state: { addMfaMethod: ["SELECT_APP"] } } } as any,
+    };
     res = {
       status: sandbox.fake(),
       redirect: sandbox.fake(() => {}),
@@ -75,175 +112,5 @@ describe("addMfaMethodPost", () => {
     const call: sinon.SinonSpyCall<Error[], unknown> = next.getCall(0);
 
     expect(call.args[0].message).to.equal("Unknown addMfaMethod: unknown");
-  });
-});
-
-describe("addMfaAppMethodGet", () => {
-  let sandbox: SinonSandbox;
-
-  beforeEach(() => {
-    sandbox = sinon.createSandbox();
-  });
-
-  afterEach(() => {
-    sandbox.restore();
-  });
-
-  it("should render the add app page", async () => {
-    const req = {
-      session: { user: { email: "test@test.com" } },
-      body: {},
-      log: { error: sinon.fake() },
-    };
-    const res = {
-      render: sinon.fake(),
-    };
-    const next = sinon.spy();
-
-    sandbox.replace(mfaModule, "generateMfaSecret", () => "A".repeat(20));
-    sandbox.replace(mfaModule, "generateQRCodeValue", () => "qrcode");
-
-    await addMfaAppMethodGet(
-      req as unknown as Request,
-      res as unknown as Response,
-      next
-    );
-
-    expect(res.render).to.have.been.calledWith("add-mfa-method/add-app.njk", {
-      authAppSecret: "A".repeat(20),
-      qrCode: await QRCode.toDataURL("qrcode"),
-      formattedSecret: "AAAA AAAA AAAA AAAA AAAA",
-      errors: {},
-      errorList: [],
-    });
-  });
-
-  it("should return an error if there is no email in the session", async () => {
-    const req = {
-      session: { user: {} },
-      log: { error: sinon.fake() },
-    };
-    const res = {};
-    const next = sinon.spy();
-
-    await addMfaAppMethodGet(
-      req as unknown as Request,
-      res as unknown as Response,
-      next
-    );
-
-    expect(next).to.have.been.calledWith(sinon.match.instanceOf(Error));
-  });
-});
-
-describe("addMfaAppMethodPost", () => {
-  let sandbox: SinonSandbox;
-
-  beforeEach(() => {
-    sandbox = sinon.createSandbox();
-  });
-
-  afterEach(() => {
-    sandbox.restore();
-  });
-
-  it("should send the request to add mfaMethod", async () => {
-    const req = {
-      body: {
-        code: "123456",
-        authAppSecret: "A".repeat(20),
-      },
-      session: {
-        id: "session_id",
-        user: { email: "test@test.com", tokens: { accessToken: "token" } },
-      },
-      log: { error: sinon.fake() },
-      ip: "127.0.0.1",
-      t: (t: string) => t,
-    };
-    const res = {
-      locals: {
-        persistentSessionId: "persistentSessionId",
-      },
-      render: sinon.fake(),
-    };
-    const next = sinon.spy();
-    const addMfaMethod = sinon.fake.returns(
-      Promise.resolve({
-        status: 200,
-        data: {
-          endPoint: "endPoint",
-          mfaIdentifier: 1,
-          methodVerified: true,
-          mfaMethodType: "AUTH_APP",
-          priorityIdentifier: "SECONDARY",
-        },
-      })
-    );
-
-    sandbox.replace(mfaModule, "verifyMfaCode", () => true);
-    sandbox.replace(
-      mfaModule,
-      "addMfaMethod",
-      addMfaMethod as typeof mfaModule.addMfaMethod
-    );
-
-    await addMfaAppMethodPost(
-      req as unknown as Request,
-      res as unknown as Response,
-      next
-    );
-
-    expect(addMfaMethod).to.have.been.calledWith({
-      email: "test@test.com",
-      otp: "123456",
-      credential: "AAAAAAAAAAAAAAAAAAAA",
-      mfaMethod: { priorityIdentifier: "SECONDARY", mfaMethodType: "AUTH_APP" },
-      accessToken: "token",
-      sourceIp: "127.0.0.1",
-      sessionId: "session_id",
-      persistentSessionId: "persistentSessionId",
-    });
-
-    expect(res.render).to.be.calledWith(
-      "common/confirmation-page/confirmation.njk",
-      {
-        heading: "pages.confirmAddMfaMethod.heading",
-        message: "pages.confirmAddMfaMethod.message",
-        backLinkText: "pages.confirmAddMfaMethod.backLinkText",
-        backLink: PATH_DATA.SECURITY.url,
-      }
-    );
-  });
-
-  it("should render an error if the code is invalid", async () => {
-    const req = {
-      body: {
-        code: "123456",
-        authAppSecret: "A".repeat(20),
-      },
-      session: {
-        id: "session_id",
-        user: { email: "test@test.com", tokens: { accessToken: "token" } },
-      },
-      log: { error: sinon.fake() },
-      ip: "127.0.0.1",
-      t: (t: string) => t,
-    };
-    const res = {
-      locals: {
-        persistentSessionId: "persistentSessionId",
-      },
-      render: sinon.fake(),
-    };
-    const next = sinon.spy();
-
-    sandbox.replace(mfaModule, "verifyMfaCode", () => false);
-
-    await addMfaAppMethodPost(
-      req as unknown as Request,
-      res as unknown as Response,
-      next
-    );
   });
 });
