+++ src/components/common/confirmation-page/confirmation.njk
@@ -1,6 +1,5 @@
 {% extends "common/layout/base-page.njk" %}
 {% from "govuk/components/panel/macro.njk" import govukPanel %}
-
 {% block pageContent %}
     {{ govukPanel({
         titleText: heading
