+++ src/components/update-confirmation/tests/update-confirmation-controller.test.ts
@@ -1,15 +1,17 @@
 import { expect } from "chai";
-import { describe } from "mocha";
+import { describe, it } from "mocha";
 
 import { sinon } from "../../../../test/utils/test-utils";
 import { Request, Response } from "express";
 
 import {
+  addMfaAppMethodConfirmationGet,
   deleteAccountConfirmationGet,
   updateEmailConfirmationGet,
   updatePasswordConfirmationGet,
   updatePhoneNumberConfirmationGet,
 } from "../update-confirmation-controller";
+import { PATH_DATA } from "../../../app.constants";
 
 describe("update confirmation controller", () => {
   let sandbox: sinon.SinonSandbox;
@@ -65,5 +67,46 @@ describe("update confirmation controller", () => {
 
       expect(res.render).to.have.calledWith("update-confirmation/index.njk");
     });
+  });
+});
+
+describe("addMfaMethodAppConfirmationGet", () => {
+  let sandbox: sinon.SinonSandbox;
+  let req: Partial<Request>;
+  let res: Partial<Response>;
+
+  beforeEach(() => {
+    sandbox = sinon.createSandbox();
+
+    req = {
+      body: {},
+      cookies: { lng: "en" },
+      i18n: { language: "en" },
+      t: (t: string) => t,
+    };
+    res = {
+      render: sandbox.fake(),
+      locals: {},
+      status: sandbox.fake(),
+    };
+  });
+
+  afterEach(() => {
+    sandbox.restore();
+  });
+
+  it("should render add mfa app confirmation page", () => {
+    addMfaAppMethodConfirmationGet(req as Request, res as Response);
+
+    expect(res.render).to.be.calledWith(
+      "common/confirmation-page/confirmation.njk",
+      {
+        pageTitleName: "pages.confirmAddMfaMethod.title",
+        heading: "pages.confirmAddMfaMethod.heading",
+        message: "pages.confirmAddMfaMethod.message",
+        backLinkText: "pages.confirmAddMfaMethod.backLinkText",
+        backLink: PATH_DATA.SECURITY.url,
+      }
+    );
   });
 });
