+++ src/locales/en/translation.json
@@ -182,6 +182,7 @@
       }
     },
     "confirmAddMfaMethod": {
+      "title": "You've added another way to get security codes",
       "heading": "You've added another way to get security codes",
       "message": "If your usual way to get security codes is not available, you can get a code from this authenticator app instead",
       "backLinkText": "Back to security"
