+++ test/unit/utils/state-machine.test.ts
@@ -23,6 +23,8 @@ describe("state-machine", () => {
       expect(nextState.events).to.all.members([
         "VALUE_UPDATED",
         "VERIFY_CODE_SENT",
+        "SELECTED_APP",
+        "SELECTED_SMS",
       ]);
     });
 
@@ -58,11 +60,45 @@ describe("state-machine", () => {
       expect(nextState.events).to.all.members([
         "VALUE_UPDATED",
         "VERIFY_CODE_SENT",
+        "SELECTED_APP",
+        "SELECTED_SMS",
       ]);
     });
 
-    it("should move state from change value state to value updated state", () => {
+    it("should move state from   change value state to value updated state", () => {
       const nextState = getNextState("CHANGE_VALUE", "VALUE_UPDATED");
+      expect(nextState.value).to.equal("CONFIRMATION");
+      expect(nextState.events).to.all.members([]);
+    });
+  });
+
+  describe("getNextState for mfa process", () => {
+    it("should move state from initial state to change value state", () => {
+      const state = getInitialState();
+      const nextState = getNextState(state.value, "AUTHENTICATED");
+      expect(nextState.value).to.equal("CHANGE_VALUE");
+      expect(nextState.events).to.all.members([
+        "VALUE_UPDATED",
+        "VERIFY_CODE_SENT",
+        "SELECTED_APP",
+        "SELECTED_SMS",
+      ]);
+    });
+
+    it("should move state from CHANGE_VALUE state to APP state when SELECTED_APP action event ", () => {
+      const nextState = getNextState("CHANGE_VALUE", "SELECTED_APP");
+      expect(nextState.value).to.equal("APP");
+      expect(nextState.events).to.all.members(["VALUE_UPDATED"]);
+    });
+
+    it("should move state from CHANGE_VALUE state to SMS state when SELECTED_SMS action event ", () => {
+      const nextState = getNextState("CHANGE_VALUE", "SELECTED_SMS");
+      expect(nextState.value).to.equal("SMS");
+      expect(nextState.events).to.all.members(["VALUE_UPDATED"]);
+    });
+
+    it("should move state from APP state to CONFIRMATION state when VALUE_UPDATED action event ", () => {
+      const nextState = getNextState("APP", "VALUE_UPDATED");
       expect(nextState.value).to.equal("CONFIRMATION");
       expect(nextState.events).to.all.members([]);
     });
